lambda/setup_okta_integration/requirements.txt
boto3 == 1.20.24
okta == 2.5.0
snowflake-connector-python == 2.7.1
